signup test

+++ code/api/Database_Subsytem/Database_Manager.py
@@ -1,3 +1,4 @@
+from asyncio.windows_events import NULL
 import mysql.connector
 
 import sys
@@ -384,12 +385,14 @@ class Database_Manager:
               "VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s)"
 
         if isinstance(academic, AcademicModel):
-
-            # to use open, we needed to save the image
-            academic.photo.save("buff.jpg")
-            with open("buff.jpg", 'rb') as file:
-                binaryData = file.read()
-            os.remove("buff.jpg")
+            if academic.photo == NULL:
+                binaryData = NULL
+            else:
+                # to use open, we needed to save the image
+                academic.photo.save("buff.jpg")
+                with open("buff.jpg", 'rb') as file:
+                    binaryData = file.read()
+                os.remove("buff.jpg")
 
             val = (academic.name, academic.surname, academic.appellation, academic.institution, academic.job_title, academic.email, academic.password, binaryData,
                    academic.research_gate_link)
